fix(securite): ferme deux acces de mutation non gardes et rend les can: explicites au niveau route

+++ app/Http/Controllers/Parametres/FeeStructureController.php
@@ -254,6 +254,8 @@ class FeeStructureController extends Controller
      */
     public function destroy(FeeStructure $feeStructure)
     {
+        abort_unless(auth()->user()->can('delete_fee_structures'), 403);
+
         $feeStructure->delete();
 
         return redirect()->route('fee-structures.index')
